Added Page Variables table
So here's the concept: What if we didn't store al the page variables in a serialized array? What if for each variable, we stored a row in page_variables? Dunno how to handle lists just yet, do we store a row for each item in the list? Dunno. But this makes the data searchable. So there's that.